refactor(platform): migrate image props getter to another file
